factor out UITestBase methods
